vcl: start migrating drawing functionality in OutputDevice into Drawables
OutputDevice must draw many different things - points, pixels, lines,
rectangles, etc. The one thing that each drawing function has is the
name Draw<Primitive> and then a bunch of parameters used to control
how the primitive will be drawn. To add a new drawing function we must
add a new Draw<NewPrimitive> function.

I am proposing a new way of achieving this - via the use of Drawables.
Each Drawable has a single execute() interface, and the constructor
takes the parameters that control how the primitive will be created.
In essence this is the command pattern as detailed by the GOF. The
advantages are that that this makes OutputDevice closed for modification
but remains extensible. It should help make testing a bit easier.

I have introduced a static convinience function Drawable::Draw() that
takes an OutputDevice and the Drawable command and executes it
immediately. Thus the syntax is not much different to what we are
using currently.

Change-Id: I74ffb609741966c7966edf73ff0ac00eede7c239