vcl: apply template pattern to Drawable::execute
Each time we draw to an OutputDevice, we follow the same basic
algorithm:

1. Add action to GDI metafile
2. Can we draw? No - exit
3. Initialize the clipping region if possible
4. Initialize the object's color
5. Initialize the object's fill color
6. Acquire the SalGraphics instance
7. Do the actual drawing!
8. Paint alpha virtual device

The only thing that varies is step 7 - the actual drawing. Thus I
now provide overridable functions that provide default behaviour,
an abstract function DrawCommand where the drawing is implemented,
and I make execute run these steps in the right order.

Now all someone who wants to implement a Drawable object has to do is
to override the functions they wish to vary, and perform their
drawing in DrawCommand. They no longer have to provide scaffolding
to make it work!

Things to note:
- each of these steps (except the actual drawing) may be optional,
  so I have put an overridable check function for each
- on some occassions we rely on the underlying command object to
  undertake all these steps and we don't want to dupliate them - so
  I provide an option to disable these scaffolding steps entirely

Change-Id: Ib9422f80a9521a3388200efd155056ab5c687c52