vcl: only enable output for necessary actions
It turns out that the very first time we check if the current item
matches an item in the member map, the output is enabled for ALL of the
actions. Really we need to disable the output at the start of the loop
and then enable output only for actions at are members of the list.

Change-Id: I3fd2b0d9ecc7fc9fe8922623169320d862aea7d3