vcl: remove MetaActionType from DrawBitmap()
This is being done so I can migrate DrawBitmap() functions to
RenderContext2. I have had to duplicate a bit of code, but have kept
this to a minimum by extracting the local function MirrorBitmap().

In the meantime, turns out that we can rename the DrawScaledBitmap() and
DrawScaledPartBitmap() functions back to just DrawBitmap().

Change-Id: Idaacc8dabd9020df702b55ad104aa9d55fcebd99